requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
google-api-core==1.31.2
google-auth==1.35.0
google-cloud-core==1.7.2
google-cloud-translate==2.0.1
googleapis-common-protos==1.53.0
grpcio==1.39.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jokeapi==1.0.2
lxml==4.6.3
MarkupSafe==2.0.1
multidict==5.1.0
packaging==21.0
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
simplejson==3.17.4
six==1.16.0
soupsieve==2.2.1
twilio==6.63.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
yarl==1.6.3
